requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.3.7
jinja2==3.1.3
python-multipart==0.0.9
